IpmiFeaturePkg/ServerManagementLib: Support Standalone MM
Split the ServerManagementLib.c into two source files:
- ServerManagementELog.c to support Event Log functions.
- ServerManagementTime.c to support GetTimeStamp function.
  - Replace gBS->Stall() with MicroSecondDelay() in TimerLib to
remove dependency on gBS.

Base Lib includes both of the two source files, StandaloneMm Lib only
includes the ServerManagementTime.c

Cc: Abner Chang <Abner.Chang@amd.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
